working single page auth
